Add the Chain Sync Handler
